requirement.txt
nibabel
pytorch_ssim
numpy
